feat(StationService) : station 삭제할 때 section에 존재하고 있다면 예외 발생

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.station.response.StationResponse;
+import wooteco.subway.exception.station.StationDeleteException;
 import wooteco.subway.exception.station.StationDuplicateException;
 import wooteco.subway.exception.station.StationNotExistException;
 
@@ -14,9 +15,11 @@ import java.util.stream.Collectors;
 public class StationService {
 
     private final StationDao stationDao;
+    private final SectionService sectionService;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionService sectionService) {
         this.stationDao = stationDao;
+        this.sectionService = sectionService;
     }
 
     public StationResponse create(String name) {
@@ -40,6 +43,9 @@ public class StationService {
     }
 
     public void deleteById(Long id) {
+        if (sectionService.existStationByStationId(id)) {
+            throw new StationDeleteException(id);
+        }
         stationDao.deleteById(id);
     }
 }
